refactor: model패키지 내부 경로 추가 생성
