+++ src/main/java/racingcargame/model/car/Car.java
@@ -1,4 +1,4 @@
-package racingcargame.model;
+package racingcargame.model.car;
 
 import racingcargame.dto.CarDto;
