+++ src/test/java/racingcargame/model/CarsTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcargame.model.car.Cars;
 
 public class CarsTest {
