+++ src/test/java/racingcargame/model/RaceCountTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcargame.model.racecount.RaceCount;
 
 public class RaceCountTest {
